dici_backend/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.66.0
matplotlib>=3.7.0
scipy>=1.11.0
